+++ tests/base_test.py
@@ -1,6 +1,6 @@
 import pytest
 
-from selene.tools import get_driver, set_driver
+from selene.browser import driver, set_driver
 from tests.acceptance.helpers.helper import get_test_driver
 
 
@@ -9,7 +9,7 @@ def setup(request):
     set_driver(get_test_driver())
 
     def teardown():
-        get_driver().quit()
+        driver().quit()
 
     request.addfinalizer(teardown)