+++ tests/examples/todomvc/pageobjects_aka_widgets_approach/todomvc_test.py
@@ -1,5 +1,5 @@
 from selene.conditions import exact_text, hidden, exact_texts
-from selene.tools import set_driver, get_driver, ss, s
+from selene.browser import set_driver, driver, ss, s
 
 from tests.acceptance.helpers.helper import get_test_driver
 from tests.acceptance.helpers.todomvc import given_active
@@ -10,7 +10,7 @@ def setup_module(m):
 
 
 def teardown_module(m):
-    get_driver().quit()
+    driver().quit()
 
 
 class Task(object):