+++ tests/integration/auto_start_test.py
@@ -1,7 +1,7 @@
 import os
 
 from selene.conditions import exact_text
-from selene.tools import visit, s, set_driver, get_driver
+from selene.browser import visit, s, set_driver, driver
 from tests.acceptance.helpers.helper import get_test_driver
 
 start_page = 'file://' + os.path.abspath(os.path.dirname(__file__)) + '/../resources/start_page.html'
@@ -28,4 +28,4 @@ def test_auto_start_2():
     s("#selene_link").should_have(exact_text("Selene site"))
 
 def teardown_module(m):
-    get_driver().quit()
+    driver().quit()