+++ tests/integration/test_scroll_to_element.py
@@ -1,7 +1,7 @@
 import os
 
 from selene import config
-from selene.tools import visit, s, get_driver
+from selene.browser import visit, s, driver
 
 start_page = 'file://' + os.path.abspath(os.path.dirname(__file__)) + '/../resources/start_page.html'
 
@@ -12,5 +12,5 @@ def setup_module(m):
 
 def test_can_scroll_to():
     visit(start_page)
-    get_driver().set_window_size(300, 400)
+    driver().set_window_size(300, 400)
     s("#invisible_link").scroll_to().click()