+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - upcoming breaking changes:
     - selene.config.app_host renamed to selene.config.base_url
       - selene.config.app_host still works but will be removed in next versions
+    - #101: deprecated selene.tools, use selene.browser instead
     
 ## 1.0.0a8 (released 16.02.2017)
   - new features added